Rename enum fields package to `enum_fields`
