+++ src/main/java/org/quiltmc/enigma_plugin/index/enum_fields/EnumFieldsIndex.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.quiltmc.enigma_plugin.index.enumfields;
+package org.quiltmc.enigma_plugin.index.enum_fields;
 
 import org.quiltmc.enigma.api.translation.representation.entry.FieldEntry;
 import org.objectweb.asm.tree.ClassNode;
